Done exercise 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -52,6 +52,17 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
     try{
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+        if (!decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+        const user = await User.findById(decodedToken.id)
+        const blog = await Blog.findById(request.params.id)
+
+        if(user._id.toString() !== blog.user.toString()){
+            return response.status(403).json({'error': 'You cannot delete others blogs'})
+        }
+
         await Blog.findByIdAndRemove(request.params.id)
         response.status(204).end()
     } catch (e) {
